Moved plots from get_data.py to app_core.py
To organize this I am separating the functions used to query the API from the usage of these functions to form plots. Now functions are in get_data.py and usage is in app_core.py. App_core imports functions from get_data like a module so they need to be in the same dir.